Added BitStream class, MDF File saving
A file BitStream.lua was created to support reading and writing binary files from the reframework\data directory
BitStream comes with most features of a BitStream class such as seek, tell, readInt, writeWString and more.
A struct-writing "REResource" class is also present, extended by specific classes for different filetypes.
EMV Engine can utilize this in the "Materials" menu, letting you input a filename of an existing MDF file (it should be the real one for your mesh) to modify and save.